json API no longer available. changed to xml

+++ db/schema.rb
@@ -2,8 +2,8 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
@@ -28,7 +28,7 @@ ActiveRecord::Schema.define(version: 2019_12_10_092007) do
   end
 
   create_table "reviews", force: :cascade do |t|
-    t.integer "latinamerica_id", null: false
+    t.bigint "latinamerica_id", null: false
     t.string "comment"
     t.integer "rating", default: 0
     t.datetime "created_at", precision: 6, null: false
